Fix NL read on Windows
In Windows, fmt::File was reading NL in text mode, because the Windows' recommended version of POSIX' open(), sopen_s(), opens in text mode by default. While in text mode, Ctrl-Z character means EOF, and the reader was looping waiting for the full file size to be read, but _read() was returning 0 bytes. Strange that this did not happen before.